Move to queries to queries folder
